+++ lib/e_menu/src/apave.tcl
@@ -7,7 +7,7 @@
 ###########################################################
 
 package require Tk
-package provide apave 4.3.0
+package provide apave 4.3.2
 
 source [file join [file dirname [info script]] apavedialog.tcl]