+++ lib/e_menu/src/pickers/color/aloupe/aloupe.tcl
@@ -11,11 +11,12 @@
 
 package require Tk
 
-package provide aloupe 1.1
+package provide aloupe 1.2
 
 namespace eval ::aloupe {
   variable solo [expr {[info exist ::argv0] && [file normalize $::argv0] eq [file normalize [info script]]}]
 }
+if {$::aloupe::solo} {wm withdraw .}
 
 # _____ Remove installed (perhaps) packages used here _____ #
 
@@ -42,21 +43,33 @@ if {$::aloupe::solo} {
   }
 }
 
-if {[catch {package require treectrl; package require Img} err]} {
-  set tclexe [auto_execok tclsh]
-  set tclrun [info nameofexecutable]
-  if {$tclexe ne {} && $tclexe ne $tclrun} {
+# ________________________ Run solo at need _________________________ #
+
+proc ::aloupe::RunSolo {} {
+  # Runs aloupe as a sole Tcl script.
+  # When aloupe runs from tclkit, it may fail. So try it with tclsh deployed.
+
+  set tclsh [auto_execok tclsh]
+  set tclexe [info nameofexecutable]
+  # tclsh may be sort of "tcl.sh" to run a tclkit
+  if {[file exists $tclsh] && [file size $tclsh]>1024 && $tclsh ne $tclexe} {
     if {$::aloupe::solo} {set aar $::argv} {set aar {}}
-    exec -- $tclexe [info script] {*}$aar &  ;# when aloupe runs from tclkit, it may fail - try to run it with tclsh deployed
+    exec -- $tclsh $::aloupe::aloupescript {*}$aar &
   } else {
-    puts "aloupe: $err"
+    puts "aloupe: $::aloupe::runerr"
   }
-  if {$::aloupe::solo} exit
+
+}
+set ::aloupe::aloupescript [info script]
+set ::aloupe::starterr [catch {package require treectrl; package require Img} ::aloupe::runerr]
+if {$::aloupe::solo && $::aloupe::starterr} {
+  ::aloupe::RunSolo
+  exit
 }
 
-::msgcat::mcload [file join [file dirname [info script]] msgs]
-
 # ________________________ Variables _________________________ #
+
+::msgcat::mcload [file join [file dirname [info script]] msgs]
 
 namespace eval ::aloupe {
   variable filename {}
@@ -605,6 +618,10 @@ proc ::aloupe::run {args} {
   # Runs the loupe.
   #  args - options of the loupe
 
+  if {$::aloupe::starterr} {
+    RunSolo
+    return
+  }
   variable my::data
   variable my::size
   variable my::zoom